Beta 2.1

+++ resources/views/servicios/index.blade.php
@@ -1,7 +1,7 @@
 <x-app-layout>
     <x-slot name="header">
         <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
-            {{ __('Citas') }}
+            {{ __('Servicios') }}
         </h2>
     </x-slot>
 
@@ -11,43 +11,34 @@
                 <div class="bg-white dark:bg-gray-800 border-b border-gray-200 dark:border-gray-700">
 
                     <div class="mb-4">
-                        <a href="{{ route('agenda.create') }}" class="bg-cyan-500 dark:bg-cyan-700 hover:bg-cyan-600 dark:hover:bg-cyan-800 text-white font-bold py-2 px-4 rounded">Agregar Cita</a>
+                        <a href="{{ route('servicios.create') }}" class="bg-cyan-500 dark:bg-cyan-700 hover:bg-cyan-600 dark:hover:bg-cyan-800 text-white font-bold py-2 px-4 rounded">Agregar Servicios</a>
                     </div>
 
                     <table class="table-auto w-full">
                         <thead>
                             <tr>
                                 <th class="px-4 py-2 text-gray-900 dark:text-white text-center">ID</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Título</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Descripción</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Estudiante</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Servicio</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Inicio</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Fin</th>
-                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Acciones</th>
+                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Nombre</th>
+                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Tipo</th>
+                                <th class="px-4 py-2 text-gray-900 dark:text-white text-center">Costo</th>
                             </tr>
                         </thead>
                         <tbody>
-                            @foreach($agenda as $agenda)
+                            @foreach($servicios as $servicios)
                             <tr>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->id }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->title }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->description }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->student->name }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->servicio->name }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->start_time }}</td>
-                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $agenda->end_time }}</td>
+                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $servicios->id }}</td>
+                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $servicios->name }}</td>
+                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $servicios->type }}</td>
+                                <td class="border px-4 py-2 text-gray-900 dark:text-white text-center">{{ $servicios->cost }}</td>
+
                                 <td class="border px-4 py-2 text-center">
                                     <div class="flex justify-center">
-                                        <a href="{{ route('agenda.show', $agenda->id) }}" class="bg-violet-500 dark:bg-violet-700 hover:bg-violet-600 dark:hover:bg-violet-800 text-white font-bold py-2 px-4 rounded mr-2">Ver</a>
-                                        <a href="{{ route('agenda.edit', $agenda->id) }}" class="bg-violet-500 dark:bg-violet-700 hover:bg-violet-600 dark:hover:bg-violet-800 text-white font-bold py-2 px-4 rounded mr-2">Editar</a>
-                                        <form action="{{ route('agenda.destroy', $agenda->id) }}" method="POST" onsubmit="return confirmDelete({{ $agenda->id }});">
-                                            @csrf
-                                            @method('DELETE')
-                                            <button type="submit" class="bg-pink-400 dark:bg-pink-600 hover:bg-pink-500 dark:hover:bg-pink-700 text-white font-bold py-2 px-4 rounded">Eliminar</button>
-                                        </form>
+                                        <a href="{{ route('servicios.edit', $servicios->id) }}" class="bg-violet-500 dark:bg-violet-700 hover:bg-violet-600 dark:hover:bg-violet-800 text-white font-bold py-2 px-4 rounded mr-2">Edit</a>
+                                        <button type="button" class="bg-pink-400 dark:bg-pink-600 hover:bg-pink-500 dark:hover:bg-pink-700 text-white font-bold py-2 px-4 rounded" onclick="confirmDelete('{{ $servicios->id }}')">Delete</button>
+
                                     </div>
                                 </td>
+
                             </tr>
                             @endforeach
                         </tbody>
@@ -60,20 +51,37 @@
 </x-app-layout>
 
 <script>
+    // forma 1
     function confirmDelete(id) {
         alertify.confirm("¿Confirm delete record?",
-            function(){
-                let form = document.createElement('form');
-                form.method = 'POST';
-                form.action = '/agenda/' + id;
-                form.innerHTML = '@csrf @method("DELETE")';
-                document.body.appendChild(form);
-                form.submit();
-                alertify.success('Ok');
-            },
-            function(){
-                alertify.error('Cancel');
-            });
-        return false; // Evita que el formulario se envíe inmediatamente
+        function(){
+            let form = document.createElement('form');
+                    form.method = 'POST';
+                    form.action = '/servicios/' + id;
+                    form.innerHTML = '@csrf @method("DELETE")';
+                    document.body.appendChild(form);
+                    form.submit();
+            alertify.success('Ok');
+        },
+        function(){
+            alertify.error('Cancel');
+        });
     }
+
+// forma 2
+/* function confirmDelete(id) {
+    alertify.confirm("¿Confirm delete record?", function (e) {
+        if (e) {
+            let form = document.createElement('form');
+            form.method = 'POST';
+            form.action = '/students/' + id;
+            form.innerHTML = '@csrf @method("DELETE")';
+            document.body.appendChild(form);
+            form.submit();
+        } else {
+            alertify.error('Cancel');
+            return false;
+        }
+    });
+} */
 </script>
